ref: extract prisma client module path normalization to a separate function
